+++ CHANGELOG.md
@@ -2,4 +2,26 @@
 
 All notable changes to Vaultsmith are documented here. Release Please maintains this file from Conventional Commits.
 
+## [0.1.1](https://github.com/Forgeplane-io/vaultsmith/compare/v0.1.0...v0.1.1) (2026-08-05)
+
+
+### Bug Fixes
+
+* **chart:** make NetworkPolicy opt-in by default ([ae3fb57](https://github.com/Forgeplane-io/vaultsmith/commit/ae3fb57c63570086c4550cbffebbffc10d91302e))
+
+## 0.1.0 (2026-08-04)
+
+
+### Features
+
+* polish Vaultsmith UI and add logo ([1aef905](https://github.com/Forgeplane-io/vaultsmith/commit/1aef9056c89b408eb4219cbda87f82776a5ed3f4))
+* polish Vaultsmith UI and add logo ([c0e30bf](https://github.com/Forgeplane-io/vaultsmith/commit/c0e30bfa6c66ad1412ae0df17ad0413aae781634))
+* **release:** prepare Vaultsmith public release ([31537c5](https://github.com/Forgeplane-io/vaultsmith/commit/31537c55c680bc4f53d5a0eda67cf3cd3b73e388))
+* sync upstream paste and deployment hardening ([5f32c0d](https://github.com/Forgeplane-io/vaultsmith/commit/5f32c0d7e00e6320dd0eeb67efa918d42eed2082))
+
+
+### Bug Fixes
+
+* harden cancellation and release CI ([ba3e43e](https://github.com/Forgeplane-io/vaultsmith/commit/ba3e43e0ea32cef76d24f0273ad2606cdc1a952c))
+
 ## [Unreleased]